update contorller logics
