--- go/components/deployment/common/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["conditions.go"],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/components/deployment/common",
-    visibility = ["//go/components/deployment:__subpackages__"],
-    deps = [
-        "//proto/api:go_default_library",
-        "//proto/api/v2:go_default_library",
-    ],
-)
