+++ go/logging/audit_log_struct.go
@@ -1,0 +1,12 @@
+package logging
+
+type AuditLogEvent struct {
+	Namespace  string
+	EntityType string
+	EntityName string
+	Procedure  string
+	Request    string
+	Response   string
+	Error      string
+	PreCRD     string
+}
